create MessageButtonsBar component
