+++ src/components/MessagesList/MessagesGroup/Message/Message.vue
@@ -112,73 +112,12 @@ the main body of the message as well as a quote.
 					</div>
 				</div>
 			</div>
-			<!-- Message Actions -->
-			<div v-if="hasActions"
-				v-show="showActions"
-				class="message__buttons-bar">
-				<Actions v-show="isReplyable">
-					<ActionButton icon="icon-reply"
-						@click.stop="handleReply">
-						{{ t('spreed', 'Reply') }}
-					</ActionButton>
-				</Actions>
-				<Actions :force-menu="true"
-					:container="container"
-					:boundaries-element="containerElement">
-					<ActionButton v-if="isPrivateReplyable"
-						icon="icon-user"
-						:close-after-click="true"
-						@click.stop="handlePrivateReply">
-						{{ t('spreed', 'Reply privately') }}
-					</ActionButton>
-					<ActionButton icon="icon-external"
-						:close-after-click="true"
-						@click.stop.prevent="handleCopyMessageLink">
-						{{ t('spreed', 'Copy message link') }}
-					</ActionButton>
-					<ActionButton :close-after-click="true"
-						@click.stop="handleMarkAsUnread">
-						<template #icon>
-							<EyeOffOutline decorative
-								title=""
-								:size="16" />
-						</template>
-						{{ t('spreed', 'Mark as unread') }}
-					</ActionButton>
-					<ActionLink v-if="linkToFile"
-						icon="icon-text"
-						:href="linkToFile">
-						{{ t('spreed', 'Go to file') }}
-					</ActionLink>
-					<ActionButton v-if="!isCurrentGuest && !isFileShare"
-						:close-after-click="true"
-						@click.stop="showForwarder = true">
-						<Share slot="icon"
-							:size="16"
-							decorative
-							title="" />
-						{{ t('spreed', 'Forward message') }}
-					</ActionButton>
-					<ActionSeparator v-if="messageActions.length > 0" />
-					<template v-for="action in messageActions">
-						<ActionButton :key="action.label"
-							:icon="action.icon"
-							:close-after-click="true"
-							@click="action.callback(messageAPIData)">
-							{{ action.label }}
-						</ActionButton>
-					</template>
-					<template v-if="isDeleteable">
-						<ActionSeparator />
-						<ActionButton icon="icon-delete"
-							:close-after-click="true"
-							@click.stop="handleDelete">
-							{{ t('spreed', 'Delete') }}
-						</ActionButton>
-					</template>
-				</Actions>
-			</div>
 		</div>
+		<MessageButtonsBar v-if="hasMessageButtonsBar"
+			v-show="showMessageButtonsBar"
+			v-bind="$props"
+			:previous-message-id="previousMessageId"
+			:participant="participant" />
 		<div v-if="isLastReadMessage"
 			v-observe-visibility="lastReadMessageVisibilityChanged">
 			<div class="new-message-marker">
@@ -192,10 +131,6 @@ the main body of the message as well as a quote.
 </template>
 
 <script>
-import ActionButton from '@nextcloud/vue/dist/Components/ActionButton'
-import ActionLink from '@nextcloud/vue/dist/Components/ActionLink'
-import Actions from '@nextcloud/vue/dist/Components/Actions'
-import ActionSeparator from '@nextcloud/vue/dist/Components/ActionSeparator'
 import Tooltip from '@nextcloud/vue/dist/Directives/Tooltip'
 import CallButton from '../../../TopBar/CallButton'
 import DeckCard from './MessagePart/DeckCard'
@@ -206,26 +141,18 @@ import RichText from '@juliushaertl/vue-richtext'
 import AlertCircle from 'vue-material-design-icons/AlertCircle'
 import Check from 'vue-material-design-icons/Check'
 import CheckAll from 'vue-material-design-icons/CheckAll'
-import EyeOffOutline from 'vue-material-design-icons/EyeOffOutline'
 import Reload from 'vue-material-design-icons/Reload'
-import Share from 'vue-material-design-icons/Share'
 import Quote from '../../../Quote'
 import isInCall from '../../../../mixins/isInCall'
 import participant from '../../../../mixins/participant'
 import { EventBus } from '../../../../services/EventBus'
 import emojiRegex from 'emoji-regex'
-import { PARTICIPANT, CONVERSATION, ATTENDEE } from '../../../../constants'
+import { CONVERSATION } from '../../../../constants'
 import moment from '@nextcloud/moment'
-import {
-	showError,
-	showSuccess,
-	showWarning,
-	TOAST_DEFAULT_TIMEOUT,
-} from '@nextcloud/dialogs'
-import { generateUrl } from '@nextcloud/router'
 import Location from './MessagePart/Location'
 import Contact from './MessagePart/Contact.vue'
 import Forwarder from './MessagePart/Forwarder'
+import MessageButtonsBar from './MessageButtonsBar/MessageButtonsBar.vue'
 
 export default {
 	name: 'Message',
@@ -235,20 +162,15 @@ export default {
 	},
 
 	components: {
-		Actions,
-		ActionButton,
-		ActionLink,
 		CallButton,
 		Quote,
 		RichText,
 		AlertCircle,
 		Check,
 		CheckAll,
-		EyeOffOutline,
 		Reload,
-		Share,
-		ActionSeparator,
 		Forwarder,
+		MessageButtonsBar,
 	},
 
 	mixins: [
@@ -388,7 +310,7 @@ export default {
 
 	data() {
 		return {
-			showActions: false,
+			showMessageButtonsBar: false,
 			// Is tall enough for both actions and date upon hovering
 			isTallEnough: false,
 			showReloadButton: false,
@@ -547,19 +469,11 @@ export default {
 				return false
 			}
 
-			return this.isSystemMessage || !this.showActions || this.isTallEnough
+			return this.isSystemMessage || !this.showMessageButtonsBar || this.isTallEnough
 		},
 
-		hasActions() {
+		hasMessageButtonsBar() {
 			return !this.isSystemMessage && !this.isTemporary
-		},
-
-		container() {
-			return this.$store.getters.getMainContainerSelector()
-		},
-
-		containerElement() {
-			return document.querySelector(this.container)
 		},
 
 		isTemporaryUpload() {
@@ -600,45 +514,6 @@ export default {
 				&& this.actorType === this.$store.getters.getActorType()
 		},
 
-		isFileShare() {
-			return this.message === '{file}' && this.messageParameters?.file
-		},
-
-		linkToFile() {
-			if (this.isFileShare) {
-				return this.messageParameters?.file?.link
-			}
-			return ''
-		},
-
-		isDeleteable() {
-			if (this.isConversationReadOnly) {
-				return false
-			}
-
-			const isObjectShare = this.message === '{object}'
-				&& this.messageParameters?.object
-
-			return (moment(this.timestamp * 1000).add(6, 'h')) > moment()
-				&& this.messageType === 'comment'
-				&& !this.isDeleting
-				&& !this.isFileShare
-				&& !isObjectShare
-				&& (this.isMyMsg
-					|| (this.conversation.type !== CONVERSATION.TYPE.ONE_TO_ONE
-						&& (this.participant.participantType === PARTICIPANT.TYPE.OWNER
-							|| this.participant.participantType === PARTICIPANT.TYPE.MODERATOR)))
-		},
-
-		isPrivateReplyable() {
-			return this.isReplyable
-				&& (this.conversation.type === CONVERSATION.TYPE.PUBLIC
-					|| this.conversation.type === CONVERSATION.TYPE.GROUP)
-				&& !this.isMyMsg
-				&& this.actorType === ATTENDEE.ACTOR_TYPE.USERS
-				&& this.$store.getters.getActorType() === ATTENDEE.ACTOR_TYPE.USERS
-		},
-
 		messageActions() {
 			return this.$store.getters.messageActions
 		},
@@ -649,10 +524,6 @@ export default {
 				metadata: this.conversation,
 				apiVersion: 'v3',
 			}
-		},
-
-		isCurrentGuest() {
-			return this.$store.getters.getActorType() === 'guests'
 		},
 	},
 
@@ -696,105 +567,22 @@ export default {
 			// again another time
 			this.$refs.message.classList.remove('highlight-animation')
 		},
+
 		handleRetry() {
 			if (this.sendingErrorCanRetry) {
 				EventBus.$emit('retry-message', this.id)
 				EventBus.$emit('focus-chat-input')
 			}
 		},
-		handleReply() {
-			this.$store.dispatch('addMessageToBeReplied', {
-				id: this.id,
-				actorId: this.actorId,
-				actorType: this.actorType,
-				actorDisplayName: this.actorDisplayName,
-				timestamp: this.timestamp,
-				systemMessage: this.systemMessage,
-				messageType: this.messageType,
-				message: this.message,
-				messageParameters: this.messageParameters,
-				token: this.token,
-			})
-			EventBus.$emit('focus-chat-input')
-		},
-		async handleDelete() {
-			this.isDeleting = true
-			try {
-				const statusCode = await this.$store.dispatch('deleteMessage', {
-					message: {
-						token: this.token,
-						id: this.id,
-					},
-					placeholder: t('spreed', 'Deleting message'),
-				})
-
-				if (statusCode === 202) {
-					showWarning(t('spreed', 'Message deleted successfully, but Matterbridge is configured and the message might already be distributed to other services'), {
-						timeout: TOAST_DEFAULT_TIMEOUT * 2,
-					})
-				} else if (statusCode === 200) {
-					showSuccess(t('spreed', 'Message deleted successfully'))
-				}
-			} catch (e) {
-				if (e?.response?.status === 400) {
-					showError(t('spreed', 'Message could not be deleted because it is too old'))
-				} else if (e?.response?.status === 405) {
-					showError(t('spreed', 'Only normal chat messages can be deleted'))
-				} else {
-					showError(t('spreed', 'An error occurred while deleting the message'))
-					console.error(e)
-				}
-				this.isDeleting = false
-				return
-			}
-
-			this.isDeleting = false
-		},
 
 		handleMouseover() {
-			this.showActions = true
+			this.showMessageButtonsBar = true
 		},
 
 		handleMouseleave() {
 			if (!this.isActionMenuOpen) {
-				this.showActions = false
+				this.showMessageButtonsBar = false
 			}
-		},
-		handleActionMenuUpdate(type) {
-			if (type === 'open') {
-				this.isActionMenuOpen = true
-			} else if (type === 'close') {
-				this.isActionMenuOpen = false
-				this.showActions = false
-			}
-		},
-		async handlePrivateReply() {
-			// open the 1:1 conversation
-			const conversation = await this.$store.dispatch('createOneToOneConversation', this.actorId)
-			this.$router.push({ name: 'conversation', params: { token: conversation.token } }).catch(err => console.debug(`Error while pushing the new conversation's route: ${err}`))
-		},
-
-		async handleCopyMessageLink() {
-			try {
-				const link = window.location.protocol + '//' + window.location.host + generateUrl('/call/' + this.token) + '#message_' + this.id
-				await this.$copyText(link)
-				showSuccess(t('spreed', 'Message link copied to clipboard.'))
-			} catch (error) {
-				console.error('Error copying link: ', error)
-				showError(t('spreed', 'The link could not be copied.'))
-			}
-		},
-
-		async handleMarkAsUnread() {
-			// update in backend + visually
-			await this.$store.dispatch('updateLastReadMessage', {
-				token: this.token,
-				id: this.previousMessageId,
-				updateVisually: true,
-			})
-
-			// reload conversation to update additional attributes that have computed values
-			await this.$store.dispatch('fetchConversation', { token: this.token })
 		},
 	},
 }
@@ -809,6 +597,10 @@ export default {
 		border-radius: 8px;
 		background-color: var(--color-background-hover);
 	}
+}
+
+.message {
+	position: relative;
 }
 
 .message-body {
@@ -936,21 +728,6 @@ export default {
 
 	&.retry-option {
 		cursor: pointer;
-	}
-}
-
-.message__buttons-bar {
-	display: flex;
-	right: 14px;
-	bottom: -4px;
-	position: absolute;
-	z-index: 100000;
-	background-color: var(--color-main-background);
-	border-radius: calc($clickable-area / 2);
-	box-shadow: 0 0 4px 0px var(--color-box-shadow);
-
-	& h6 {
-		margin-left: auto;
 	}
 }
 </style>
